Creation of rules on files for non-determinitic RSEs break the conveyor : Closes #36
